requirements.txt
beautifulsoup4==4.3.2
python-dateutil==2.1
